addin part 4.21

+++ part4/bloglist/src/controllers/blogs.js
@@ -48,7 +48,16 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
   const blog = await Blog.findById(request.params.id)
+  if(blog.user.toString() !== user._id.toString()){
+    return response.status(401).json({errır: 'can not be deleted other than the creator'})
+  }
+
   if(blog){
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
